+++ components/src/list/TriggerList.ts
@@ -174,7 +174,7 @@ export class TriggerList extends ContentList<Trigger> {
 
   constructor() {
     super();
-    this.valueKey = 'id';
+    this.valueKey = 'uuid';
     this.bulkActions = [
       { key: 'archive', label: 'Archive', icon: Icon.archive }
     ];